+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Client Versions ##
+* v9.45.0 (2014-10-12)
+	* add dividend fee of 0.0002 XCP per recipient (protocol change: 328000)
 * v9.44.0 (2014-09-22)
 	* server action requires `server` positional argument
 	* lockfile
 	* made `--force` server‐side only, moved after `server` argument
-	* multiple sources, destinations (testnet protocol change: 281000)
-	* multi‐signature support (testnet protocol change: 281000)
+	* multiple sources, destinations (testnet protocol change: 303000)
+	* multi‐signature support (testnet protocol change: 303000)
 * v9.43.0 (2014-09-14)
 	* generate movements hash for each block (start at block: 322000)
 * v9.42.0 (2014-09-04)